final changes with pagination

+++ src/app/views/manage/author/authors.component.ts
@@ -4,7 +4,7 @@ import { AuthorService } from "./crude/author.service";
 import { Author } from "./crude/author";
 import { Router } from '@angular/router';
 import { Product } from '../product/crude/product';
-
+import { environment } from '../../../../environments/environment';
 
 @Component({
   templateUrl: 'authors.component.html'
@@ -13,11 +13,13 @@ export class AuthorsComponent implements OnInit {
   
   authors: Observable<Author[]>;
   author: Author = new Author();
+  DispName : string;
 
   constructor(private authorService: AuthorService,
     private router: Router) {}
 
   ngOnInit() {
+    this.DispName = environment.ProjectName;
     if(localStorage.getItem("userName")=="" || localStorage.getItem("userName")==null){
       this.router.navigate(['/login']);
     }
@@ -29,13 +31,15 @@ export class AuthorsComponent implements OnInit {
   }
 
   deleteAuthor(id: number) {
-    this.authorService.deleteAuthor(id)
-      .subscribe(
-        data => {
-          console.log(data);
-          this.reloadData();
-        },
-        error => console.log(error));
+    if(confirm("Are you sure to delete ")) {
+      this.authorService.deleteAuthor(id)
+        .subscribe(
+          data => {
+            console.log(data);
+            this.reloadData();
+          },
+          error => console.log(error));
+      }    
   }
 
   updateAuthorStatus(id: number){
