+++ src/app/views/manage/producttab/producttab.component.ts
@@ -4,7 +4,7 @@ import { ProductTabService } from "./crude/producttab.service";
 import { ProductTab } from "./crude/producttab";
 import { Product } from "../product/crude/product";
 import { Router } from '@angular/router';
-
+import { environment } from '../../../../environments/environment';
 
 @Component({
   templateUrl: 'producttab.component.html'
@@ -16,35 +16,51 @@ export class ProductComponent  implements OnInit {
   producttabs: Observable<ProductTab[]>;
   product: ProductTab = new ProductTab();
   products: Observable<Product[]>;
-  
+  projectName : string;
+  producttab : ProductTab = new ProductTab;
 
   constructor(private productTabService: ProductTabService,
     private router: Router) {}
 
+    
+
   ngOnInit() {
+    this.projectName = environment.ProjectName;
     if(localStorage.getItem("userName")=="" || localStorage.getItem("userName")==null){
       this.router.navigate(['/login']);
     }
     this.reloadData();
+   
   }
  
   reloadData() {
     this.products = this.productTabService.getProduct();
+    
+    if(localStorage.getItem("productId")!=null){
+      console.log("This is calll---productId-------------------"+localStorage.getItem("productId")) 
+     this.producttab.productId = Number(localStorage.getItem("productId"));
+     this.producttabs = this.productTabService.getProductTabList(this.producttab.productId);
+   }else{
+    this.producttab.productId = 0;
+   }
   }
 
   onOptionsSelected(value:number){
     console.log("the selected value is " + value);
+    localStorage.setItem("productId",value+"");
     this.producttabs = this.productTabService.getProductTabList(value);
 }
 
   deleteProduct(id: number) {
-    this.productTabService.deleteProductTab(id)
+    if(confirm("Are you sure to delete ")) {
+       this.productTabService.deleteProductTab(id)
       .subscribe(
         data => {
           console.log(data);
           this.reloadData();
         },
         error => console.log(error));
+    }   
   }
 
   updateProductStatus(id: number){
